+++ packages/multichain/src/caip25Permission.ts
@@ -11,7 +11,7 @@ import {
   CaveatMutatorOperation,
   PermissionType,
 } from '@metamask/permission-controller';
-import type { CaipAccountId, Json } from '@metamask/utils';
+import type { CaipAccountId, CaipChainId, Json } from '@metamask/utils';
 import {
   hasProperty,
   KnownCaipNamespace,
@@ -21,9 +21,8 @@ import {
 } from '@metamask/utils';
 import { cloneDeep, isEqual } from 'lodash';
 
-import { getEthAccounts } from './adapters/caip-permission-adapter-eth-accounts';
 import { assertIsInternalScopesObject } from './scope/assert';
-import { isSupportedScopeString } from './scope/supported';
+import { isSupportedAccount, isSupportedScopeString } from './scope/supported';
 import { mergeInternalScopes } from './scope/transform';
 import {
   parseScopeString,
@@ -56,6 +55,7 @@ export const Caip25EndowmentPermissionName = 'endowment:caip25';
 
 /**
  * Creates a CAIP-25 permission caveat.
+ *
  * @param value - The CAIP-25 permission caveat value.
  * @returns The CAIP-25 permission caveat (now including the type).
  */
@@ -68,7 +68,9 @@ export const createCaip25Caveat = (value: Caip25CaveatValue) => {
 
 type Caip25EndowmentCaveatSpecificationBuilderOptions = {
   findNetworkClientIdByChainId: (chainId: Hex) => NetworkClientId;
-  listAccounts: () => { address: Hex }[];
+  listAccounts: () => { type: string; address: Hex }[];
+  isNonEvmScopeSupported: (scope: CaipChainId) => boolean;
+  getNonEvmAccountAddresses: (scope: CaipChainId) => string[];
 };
 
 /**
@@ -113,17 +115,45 @@ export function diffScopesForCaip25CaveatValue(
 }
 
 /**
+ * Checks if every account in the given scopes object is supported.
+ *
+ * @param scopesObject - The scopes object to iterate over.
+ * @param listAccounts - The hook for getting internalAccount objects for all evm accounts.
+ * @param getNonEvmAccountAddresses - The hook that returns the supported CAIP-10 account addresses for a non EVM scope.
+ * addresses.
+ * @returns True if every account in the scopes object is supported, false otherwise.
+ */
+function isEveryAccountInScopesObjectSupported(
+  scopesObject: InternalScopesObject,
+  listAccounts: () => { type: string; address: Hex }[],
+  getNonEvmAccountAddresses: (scope: CaipChainId) => string[],
+) {
+  return Object.values(scopesObject).every((scopeObject) =>
+    scopeObject.accounts.every((account) =>
+      isSupportedAccount(account, {
+        getEvmInternalAccounts: listAccounts,
+        getNonEvmAccountAddresses,
+      }),
+    ),
+  );
+}
+
+/**
  * Helper that returns a `authorizedScopes` CAIP-25 caveat specification
  * that can be passed into the PermissionController constructor.
  *
  * @param options - The specification builder options.
  * @param options.findNetworkClientIdByChainId - The hook for getting the networkClientId that serves a chainId.
  * @param options.listAccounts - The hook for getting internalAccount objects for all evm accounts.
+ * @param options.isNonEvmScopeSupported - The hook that determines if an non EVM scopeString is supported.
+ * @param options.getNonEvmAccountAddresses - The hook that returns the supported CAIP-10 account addresses for a non EVM scope.
  * @returns The specification for the `caip25` caveat.
  */
 export const caip25CaveatBuilder = ({
   findNetworkClientIdByChainId,
   listAccounts,
+  isNonEvmScopeSupported,
+  getNonEvmAccountAddresses,
 }: Caip25EndowmentCaveatSpecificationBuilderOptions): EndowmentCaveatSpecificationConstraint &
   Required<
     Pick<EndowmentCaveatSpecificationConstraint, 'validator' | 'merger'>
@@ -152,7 +182,7 @@ export const caip25CaveatBuilder = ({
       assertIsInternalScopesObject(requiredScopes);
       assertIsInternalScopesObject(optionalScopes);
 
-      const isChainIdSupported = (chainId: Hex) => {
+      const isEvmChainIdSupported = (chainId: Hex) => {
         try {
           findNetworkClientIdByChainId(chainId);
           return true;
@@ -163,11 +193,17 @@ export const caip25CaveatBuilder = ({
 
       const allRequiredScopesSupported = Object.keys(requiredScopes).every(
         (scopeString) =>
-          isSupportedScopeString(scopeString, isChainIdSupported),
+          isSupportedScopeString(scopeString, {
+            isEvmChainIdSupported,
+            isNonEvmScopeSupported,
+          }),
       );
       const allOptionalScopesSupported = Object.keys(optionalScopes).every(
         (scopeString) =>
-          isSupportedScopeString(scopeString, isChainIdSupported),
+          isSupportedScopeString(scopeString, {
+            isEvmChainIdSupported,
+            isNonEvmScopeSupported,
+          }),
       );
       if (!allRequiredScopesSupported || !allOptionalScopesSupported) {
         throw new Error(
@@ -175,22 +211,21 @@ export const caip25CaveatBuilder = ({
         );
       }
 
-      // Fetch EVM accounts from native wallet keyring
-      // These addresses are lowercased already
-      const existingEvmAddresses = listAccounts().map(
-        (account) => account.address,
-      );
-      const ethAccounts = getEthAccounts({
-        requiredScopes,
-        optionalScopes,
-      }).map((address) => address.toLowerCase() as Hex);
-
-      const allEthAccountsSupported = ethAccounts.every((address) =>
-        existingEvmAddresses.includes(address),
-      );
-      if (!allEthAccountsSupported) {
+      const allRequiredAccountsSupported =
+        isEveryAccountInScopesObjectSupported(
+          requiredScopes,
+          listAccounts,
+          getNonEvmAccountAddresses,
+        );
+      const allOptionalAccountsSupported =
+        isEveryAccountInScopesObjectSupported(
+          optionalScopes,
+          listAccounts,
+          getNonEvmAccountAddresses,
+        );
+      if (!allRequiredAccountsSupported || !allOptionalAccountsSupported) {
         throw new Error(
-          `${Caip25EndowmentPermissionName} error: Received eip155 account value(s) for caveat of type "${Caip25CaveatType}" that were not found in the wallet keyring.`,
+          `${Caip25EndowmentPermissionName} error: Received account value(s) for caveat of type "${Caip25CaveatType}" that are not supported by the wallet.`,
         );
       }
     },