+++ packages/name-controller/src/NameController.test.ts
@@ -621,7 +621,7 @@ describe('NameController', () => {
             variation,
           } as SetNameRequest),
         ).toThrow(
-          `Must specify a chain ID in hexidecimal format or the fallback, "*", for variation when using 'ethereumAddress' type.`,
+          `Must specify a chain ID in hexadecimal format or the fallback, "*", for variation when using 'ethereumAddress' type.`,
         );
       });
 
@@ -2015,7 +2015,7 @@ describe('NameController', () => {
               // eslint-disable-next-line @typescript-eslint/no-explicit-any
             } as any),
           ).rejects.toThrow(
-            `Must specify a chain ID in hexidecimal format or the fallback, "*", for variation when using 'ethereumAddress' type.`,
+            `Must specify a chain ID in hexadecimal format or the fallback, "*", for variation when using 'ethereumAddress' type.`,
           );
         },
       );